Add configuration for a cdh5.10.0 VM
Getting pyspark all setup for local development can be a bit of a pain,
not only for getting pyspark but because the pyspark 1.6.0 we run in
production doesn't quite match the 1.6.0 you can download from apache.
To easy that burden this sets up an ubuntu VM with vagrant, installs
the same version of CDH we use in prod, and installs the necessary
python dependencies to get everything going.