Setup tox for running flake8 and pytest
More plumbing for general python project setup. This allows the tox
command to be used to run all the tests, syntax checking, etc that we
want to run on every commit. Work to get this running as part of jenkins
CI pipeline will be further along. One of the constraints there will be
getting the cdh5.10.0 spark packages installed, but shouldn't be too
difficult.

* Moves the virtualenv for running tox in the vm to /vagrant/venv to keep
the mess in one place. Tried to avoid needing an extra virtualenv, as
tox builds venvs anyways, but tox+pip weren't playing nice and errored
out with the .[test] dep otherwise.
* Switched to debian jessie. Prod is moving that direction, and it's no
harm to switch now before anything complex is setup.
* replace requirements.txt with setup.py
* Add a LICENSE file, it's MIT.
* Adjust the Vagrantfile to use NFS share. With the default share
tox/virtualenv were unable to create hardlinks.

Change-Id: Id57bd5fd0476fc061d4b0a1cd93a1b2f639b7ed4