Example script of full training pipeline
This script gives a general overview of how the pieces of the
mjolnir library work together. It might also provide some area
to bikeshed about what the api's should look like.

In the long run there will be some more advanced version of this script
with various parameters for doing feature engineering, and there will
be python scripts in the discovery analytics repository called by oozie
that run pieces of the pipeline saving intermediate stages that are
potentially useful for feature engineering, and automating the process
of building a model from beginning to end.

Change-Id: If5e06ad51408687ccac9491edc7ea91ca2d57cf7