Calculate NDCG of input click data
Adds a step to the input data pipeline to calculate weighted
and unweighted ndcg of the incoming click data. This allows
us to compare the quality of a model trained with this data
against the historical search results.

The weighted ndcg calculates the ndcg for each individual user session
in the input data. The unweighted ndcg calculates the ndcg only once
per query, using the average position a hit was shown to sessions at.
The unweighted value should be the most directly comparable to xgboost
training results, as we are not (yet) able to provide weights to
xgboost4j-spark.

Change-Id: Ic28b6c345c656f670fe5f954d0fe520b93352369