Better grouping of normalized queries
After reviewing numerous query groupings, it was clear that the lucene
stemmer does a pretty good job in most cases, but has edge cases that
groups completely unrelated queries together. The most obvious is
perhaps when lucene returns an empty string due to stop words (to be or
not to be), but there are various other cases as well.

Add another step after the lucene stemmer which compares the similarity
of result sets returned for queries. This uses a relatively simple
jaccard binary simliarity measure to compare the sets of returned page
ids, and then a very simple clustering algorithm to group together
only queries that have a minimum threshold of result similarity.

Removes the norm_query field and replaces it with norm_query_id. This
new field is unique across all wikis and can be grouped on directly.
Likely though the need to keep `wikiid` in the data frame for later
processing will mean most actions still group on (wikiid,
norm_query_id).

The current threshold for clustering of 0.5 is arbitrarily chosen, it
essentially says if we have 20 results displayed for each query then
at least 14 results must be shared between the two. It seems plausible
this could go down to 0.33, which is an overlap of 10, but it needs
to be more thoroughly reviewed.

Change-Id: I717c97cebf1762071c4e7bdf762aebb3a4520a97