Add feature collection over kafka
Adds a new option to data_pipeline to collect features via kafka,
and a new cli script kafka_daemon.py that reads requests out of
kafka, runs them against localhost, and sends the results back
into kafka. This is necessary because there is no firewall opening
for analytics to talk to relforge, we need to proxy over an existing
firewall hole.

This is probably a bit fragile still, but seems to work. This will only
be used for testing new query types that require a plugin update to
be deployed to relforge, models built for production usage will still
need to collect features from the production clusters. If we lose a
small % of queries when testing new queries it's not a big deal.

This works for an arbitrary number of kafka partitions in all topics. It
also works with an arbitrary number of daemons running.  Note though
that only as many daemons will do work as partitions in mjolnir_request
topic.

The implementation is as follows:
* The spark driver takes note of the current offset of all partitions
 of the 'mjolnir_response' topic.

* Elasticsearch msearch requests are generated by spark workers and
 pushed into a kafka topic called 'mjolnir_request'

* When request generation is complete a special 'end run' message is
 sent to all partitions of the 'mjolnir_request' topic.

* The daemon(s) running on relforge read requests from
 'mjolnir_request', run them, and sends the results back in the
 'mjolnir_result' topic.

* When the daemon sees the special 'end run' message it stops consuming
 messages and waits for the internal queue to drain. Once it has
 drained it sends back a message in a special 'mjolnir_complete' topic
 that contains the maximum offset in all partitions in
 'mjolnir_result'. This is the maximum possible offset that a result for
 a query for that run from the partition being read could be stored.

* The spark driver instance watches the 'mjolnir_complete' for all
 of the 'end run' messages to be reflected back and takes note of
 the maximum reported offset for all 'mjolnir_response' partitions.

* A spark rdd is created to read all kafka messages between the offset
 recorded in the first step and the maximum offset reported in the
 'mjolnir_complete' topic. This should encompase all responses to
 the requests.

Bug: T162072
Bug: T162059
Change-Id: Iad0be3fe95a5e10ba0d4ed3335aac1fca2dc640c