Vectorize jaccard similarity
The calculation of the similarity matrix dominates the time taken to
group together queries, at about 200ms for 175 queries and 6.5s for 1000
queries.  In the actual data our largest group is those containing only
stop words at ~550 unique queries. This patch brings down the execution
time on 1000 queries to around 300ms, and 175 queries down to 6ms (on my
laptop). This amounts to a 20-30x speedup vs the naive version.

Change-Id: I7d40ee747b089bfeb4851aab05ce34680d9392ac