Setup python logging
Initializes python logging in the shell scripts so we can get useful
debugging information. This is particularly useful when debugging kafka
to see what the library is doing.

This is a bit of duplicate code, maybe some refactoring needs to be done
to have a single entry point, or some sort of shared entry point code
being called, but for now this probably works fine.

Change-Id: I2e4200f6e12e77e4ffb230280de3eefe1f9b3f20