Extract feature vectors from the ltr plugin
Allows to use features defined in the plugin store to record feature
vectors.
Not a big fan of the way I integrated this feature, it certainly needs
some cleanups to reuse a bit more code.
Also I thought that I won't use _msearch but I found that it allowed
me to reuse a bit more code...

Bug: T168813
Change-Id: Ic258f6bee8bcb0cee3ab802935fc6ab404c85c25