Calculate CV splits on the driver
The calculation of CV splits is only deterministic if it is
called each time with the exact same rows per partition. This
guarnatee is not met because catalyst decides how many partitions
a parquet df read from disk should have based on the query plan. This
led to intermittent information leakage between test and train, and
ocasionally the groupData passed to xgboost being incorrect if the
splits were re-calculated after calculating the groupData.

This new solution collects all the rows to the driver and calculates
the splits there. This guarantees the splits are calculated a single
time and any re-calculation will use the provided splits.

Change-Id: I8a8189f999d5a59cbf699b39b98b4e3a4ce0549b