Use a custom Accumulator to collect feature names
and use lexical order for feature ordinals.
This a preliminary patch to allow mjolnir to work on an unknown
feature set.
Currently mjolnir owns the list of features and knows how to map
their names with an ordinal in the feature vector used for training.
Use a custom accumulator so that feature names can be discovered while
logging them. The strategy here is to always record feature values
in alphabetical order.

Use nolock as a nfs mount option (it solved by lock issue with sqlite3)

Bug: T168813
Change-Id: I1758166c9cf746a27626216b6018b476b3a841df