Take the desired number of samples in data_pipeline.py
The sampling rate in the data pipeline was done by having the user
provide the number of normalized queries to work with, but this
isn't a value the user ever cares about. What they actually care
about is how many rows are in the output data per-wiki. This calculation
is very approximate, but hopefully it is close enough to be
useful.

Testing on a fake dataset of 5M sessions, this has a max absolute error
rate (on 100 repetitions) of around 20% at 10k samples, 4% at 100k
samples and 0.5% at 1M samples.  Against the full prod dataset with
90M sessions and 500M unique (query, hit_page_id) pairs a samples of
1M queries had a max absolute error of 2% which seems acceptable.

Change-Id: Ic4f1abc9e33cdc9c3d7bc346355a750c8ce1ef68