Unbreak tune_results serialization
If0a17907b moved tune_results serialization later in the process, which
exposed that the group data for training the main model was added to
the object to be serialized. Since this is a java object it cannot be
serialized and break the training script. Fix by making a copy of
parameters to do the final model training with.

Change-Id: I639bf507c556222efdd70dda908204a6faa1db23