Setup tuning profiles for different size datasets
The appropriate hyperparameter space to search depends both
on the distribution of data in the dataset being trained against,
as well as the size of the dataset itself. For example a depth of 10
makes no sense when training a model with only 200k observations.

Setup some profiles based on size and set their parameters based on
the recent round of training models for 20 different wikis.

Change-Id: I1f2fe53f1080a6af1c7d71cb1d6d27693bd131a4