Some scripts to automate training and model upload
Add a script above the current cli scripts to automate the configuration
we need to train on many different wikis. Uses a YAML config file to
know what to do for various wikis.

I might have went too far with the configuration, now pretty much
everything to do for commands, except the bare bones names of commands
to be executed, comes from the config file. This does make it pretty
flexible and made adding new commands (such as pyspark) relatively
easy, but it might have just shifted the difficulties elsewhere...

Upload script could be improved, but should work as a first
implementation.

Change-Id: I19e02ac51b155eba7b9e296d203e03e7499a996d