respect XDG_CACHE_HOME for ivy cache in test
By default spark jar resolution uses ivy and defaults
the cache to the users home directory. In the docker
CI pipeline the user has no home directory and this
fails. Additionally CI wants to keep all caches in
a single directory to make it easy to re-use in a
standardized method of sharing caches between tests.

Change-Id: I597bdab6b58ca66444bb447ecfb9dd788a46a3fd