Make spark utility working dir configurable
To support deployment to the analytics cluster via scap3 we
need to be able to explicitly set the working dir, rather than
depending on it being auto detected (incorrectly). Split it
out into its own top-level parameter.

This has one important change in particular, example_train.yaml really
is just an example now. At a bare minimum the working_dir needs to
either be removed or updated to point to an appropriate location. This
should be a non-issue with the new mjolnir-deploy repository setting
up mjolnir in the analytics cluster and having appropriate configuration
files.

Change-Id: I2a2d3e899075837f2d9e25698f495e0165a76c4c