Port DBN from clickmodels to scala
DBN takes quite some time when running against a full run of wikis. On
one run that resulted in 90M observations across 20 wikis it took over
20 minutes.  In the naive benchmark (included) this scala version is >
20x faster than the python implementation, and is perhaps easier to
follow since we remove unused functionality ( the intents and layouts).

This does not yet include the python side of calling this, because the
way the CI works we will need to publish a new mjolnir jar with this
code and wanted to let it get reviewed first.

For performance reasons this almost exclusively uses arrays, and most
inner loops are using while instead of more idiomatic map or fold. This
conversion gave an ~3x speedup, which seems worthwhile. This is very
allocation heavy, over 1GB/sec in my tests, but optimizing out the
allocations seemed like a big pain.

Change-Id: I7231590a18b7f8fe2552997bc4c702ee635d06e5