Add option to train using external memory
I'm not sure what exactly has changed, but i'm unable to complete a full
round of training on wikis with large (~35M) numbers of observations
keep getting killed by spark. I tried increasing memory overhead from 9G
to 12G but it still keeps dieing. I'm wary of allocating even more
memory than that, as we are asking for a significant % of cluster
memory.

Take advantage of xgboost's external memory implementation to prevent
the memory explosion. This basically writes out the features matrix to
disk and memory maps it, depending on the kernel disk cache to keep it
in memory where possible. This is likely a little slower, but still
faster than killing executors and regularly restarting training.

Additionally while testing this out i noticed that while some small
number of executors are killed for overrunning yarns 12G of allowed
memory, other are getting OOM's from java. Give the java tasks another
1G of memory. This brings the memory to cpu ratio of the large tasks
to 2.22 GB/core.  The cluster itself has 2.26GB/core so this should
be acceptable. This change will also need to be applied to the spark
config in the mjolnir/deploy repo.

Change-Id: Ie283c1c58d8395054164f1c0157e1a709d1cccc4