Allow spark to keep the full data pipeline in memory
Something with the upgrade to spark 2.1.2 has caused us to
recompute lots of data over and over again in the data pipeline.
This is particularly egregious for normalization and feature collection
steps which take an hour each on a full run of data.

I tested out simply not unpersisting our data and everything seems to
work fine. We have ~100G of memory available for caching and only
end up using 50G by not unpersisting. Figuring out what data is
available is also much easier if we don't have to think about when
to unpersist what data.

Change-Id: Iedf259e481055444f369c528a56bee372e57595e