Remove support for collecting results from elastic queries
This support was useful before the ltr plugin had feature
logging capabilities, but now that that is working well
we need to standardize on the single use case rather than
having multiple ways of collecting data.

Individual queries that are used as features have been split
into a featuresets.py file. A followup patch will add support
for managing the featuresets we use.

Change-Id: I3d821468aa3804f14e392ab777b778a851f88602