XGBoostModel.eval needs same shape features as training
The txt file emitted by DataWriter has to be read by
lightgbm and xgboost. As such it starts features at idx 1 to make
lightgbm happy (which stores the label at idx 0). This broke
XGBoostModel.eval because it was not providing the empty feature at
index 0 that training sees.

While just a hack, this injects an empty feature at index 0 so
the evaluation see's the same shape of data as training.

Change-Id: I78a0de3790cafdd58cbc9b4549ff750489923e77