Reorder dmatrix creation and rabit init
When an executor tries to perform a second round of training it is
unable to create DMatrix instances, reporting that rabit has been
shutdown. To get around this we need to initialize Rabit before
creating any dmatrixes. This is problematic because xgboost only loads
binary data files for non-distributed training. The fix for that is a
related patch, I5dfad014. The two patches combined allow distributed
training to run on yarn.

Relatedly, this means that executors that do distributed training
cannot later do single instance training. We don't currently trigger
this, but it is something that might bite us in the future.

Related: I5dfad014ae4e0fcf3d4d293ae971bdc7511c65cc
Change-Id: I506a9fb9b20a3d25f7df20188097da4fe1c3a4cb