Feature selection
Implement feature selection in spark. Defaults to 50 features
and the mrmr algorithm which performed well in offline testing.
Brings in a new spark dependency to do the actual selection.

QuantileDiscretizer, rather than the discretizer provided, was
used for performance considerations. QuantileDiscretizer runs
on datasets of 35M rows in a few minutes, where the one provided
took hours and was generally inefficient. The feature selection
algo is also somewhat inefficient (in terms of idle spark executors)
but completes in a reasonable time span (75M obs in 19 wikis
runs in under an hour).

Change-Id: Ib698073f50e8feefe05f4e1ef0c73d39ff8d0b5c