Daemon for bulk loading data into es from kafka
Support for loading features, such as popularity score, calculated
in the analytics cluster into the production elasticsearch clusters.

This specifies a jsonschema for the messages that are accepted. Each
message must contain an index, an id, and a list of properties to be
updated. The daemon only accepts updates to properties it has been
configured to apply.

For this initial use case topic writes will generally be expected from
an oozie job in the `wikimedia/discovery/analytics` repository called
transferToKafka.py.

Supports writing to multiple clusters configured by cli flags. The
configured clusters are queried to determine where to direct
individual write requests. If an index of the same name exists on
multiple clusters only one will be written to.

Change-Id: I924f7dd94b8c3f4fc6dbcb718379b06cf852dc78