Introduce metrics collection via prometheus
With the daemons deployed for production use we need observability
into what is going on. Prometheus seems to be the way forward for
prod service.

Reports metrics about result of processing each incoming request,
the total rate of processing, and the time it takes to ship each
batch to elasticsearch.

Bug: T201948
Change-Id: I1f952f592c9c84864aeadf04873ab8f15769b409