Try to make xgboost stop killing test suite
My latest theory is that multiple xgboost train runs are running in
parallel on the same JVM and causing our intermittent failure. While
running tests wrap the train method with a mutex so we guarante a
single runner at a time. I was experiencing the issue intermittently
and with this fix ran 50 iterations of test_integration.py without
fail.

Change-Id: I81678a718b74ba5f4c6854791a089b23e4e596d8