Toggle msearch daemon on/off based on server load
Queries elasticsearch for the number of queries that have been issued
against a canary index, probably enwiki_content. Only enable kafka
subscription when the query rate to canary index is below some
threshold. If we are subscribed and query rate rises above the
threshold the consumer exits and goes to a waiting state.

Uses an EMA with alpha set to smooth qps over ~60 seconds worth of
data. Pings elasticsearch up to every 15 seconds to keep filling EMA
with new data. This needs to be relatively short while generating
load so we stop quickly if prod load arrives to the cluster.

To avoid making hundreds of unnecessary requests per minute for stats
this uses a heuristic where each time stats are fetched the interval
Change-Id: I6df613c9806c02017831cd1977b76b11d1e72357