Use kafka 0.10+ timestamp support for msearch
With kafka 0.9 to figure out where we needed to read in the topic we
needed to do some funny offset handling. With 0.10+ kafka now offers
timestamp based offsets, so all we need to do is know when we started
producing messages and when the last end run sigil comes back. This
doesn't remove too much code, but simplifies the mental model.

Change-Id: Idc2fd66425a11fee06dccfd31430109fa3c7e07c