Convert msearch daemon to polling
The next patch which toggles the daemon on/off based on cluster
load needs a way to exit the consuming state even when there
are no records coming in. That requires a switch to the polling
api so we have a regular opportunity to make a decision instead
of waiting for the next message.

By separating reading the kafka topic a bit more from
the rest we are able to more easily add some tests of messages
flowing through. Also noticed while working with this we pass
the kafka record around everywhere, but all we ever use is the
deserialized message it contains. Pass that around instead.

Change-Id: I8f9d8823ee8a24eb67b42efb2ac8200420f41aa4