Provide valid float to msearch daemon EMA metric
Metrics need to be provided with valid float's or they generate
exceptions every time they are pinged. The EMA was returning None
before it was initialized on startup causing unnecessary exceptions
to be logged. Instead default the metric to 'nan' which is a valid
float accepted by python and prometheus.

Change-Id: I83138075a2e2b8ae572f3ed894ed15b4056f4c8b