Convert KafkaRDD usage to kafka-python
The KafkaRDD support in pyspark only supports the kafka protocol as of
0.8. Communicating with newer kafka servers using this older protocol
induces additional load on the servers and can result in OOM's inside
the servers.

Since the rest of mjolnir is using kafka-python library,
switch this usage to kafka-python as well. The usage is very straight
forward, consuming the specified partitions through the chosen offset
ranges and emitting the records.

Bug: T219932
Change-Id: I36d675660335067f731d630b55d3963c5aee1945