Implement daemon for handling bulk import via swift
Convert bulk daemon to work off of files uploaded to swift and
notified over kafka, rather than pushing the whole update over kafka.
This moves the handling of the writing the updates to the analytics
side, with the daemon now expecting files in the elasticsearch bulk
export format.

Tries to remain flexible to handle future ideas such as importing
glent indices and LTR models within the same framework of swift+kafka.

Bug: T227364
Change-Id: Idab5f8d9249bc61adbdeefc28166cd6cd680bb77