Import glent suggestions over swift
Configure a second swift container for the bulk daemon
and implement a handler for the glent uploads. This expects
glent to ship a single message with an object_prefix that
will yield all relevant files. These files must be in elasticsearch
bulk import format and must not specifiy and index name or doc type.

Daemon manages the full "standard" lifecycle and deletes indices as
they become unused. Maintains an alias for most recent deployed
version, and as such is always deleting 2 indices ago after an import.
Rollback is possible but must be performed externally by removing
the current prod aliases and replacing them with the rollback aliases.

Bug: T227364
Change-Id: Id94f18d1c06a580eee24721422ff9a2dedcdf7f3