Introduce basic Dockerfile
Simply installs mjolnir into the container, lacking things
like spark.  Intended for testing the daemons which need to
access a variety of external services for integration testing.

Does not run anything by default, mjolnir-utilities.py is on
the PATH and can be provided CLI options as necessary.

Adds a requirements.txt file to speed along the container
build. This should not be considered authoritative, rather
an optimization so the mjolnir code can be updated without
reinstalling all the dependencies.

Change-Id: Ic3cca6e43bc8fcebeac9497f4b2d06bb9351e382