Add progressive typing to daemons
Doesn't go too far, but gets a plausible foundation in place for using
type annotations with mjolnir. Many other parts of mjolnir are
documented with types, but they are in doc blocks and not processed.

Due to having to support python 3.5 this is peppered with typing.cast
calls, particularly for things like initializing an empty container.
This can be switched to property annotations once we abandon 3.5.

Overall tries to not change the implementation of anything, it mostly
adds types. There was a bit of code in the msearch daemon that mixed
ints and floats, it's all floats now.

Quite a few things are only typed `Mapping`. This is basically saying
an immutable dict of something, and over time perhaps many of these
uses can be replaced with NamedTuple. For now accept that the
type checking will be limited when dealing with most dictionary
return types.

Change-Id: I071b92ba4bb78937ab77e2ee36633906f93661ae