Update feature selection to use QuantileDiscretizer
In prior versions of spark (now using 2.3.1) the QuantileDiscretizer
could only do one column at a time. Now that it can do all of the
columns in a single pass switch over to that and get rid of the thread
pool we were using to run lots of them in parallel.

This additionally means we no longer need spark-MDLP-discretization.

Change-Id: Ibe8b27e82f75276649f3c07eac471342fa757d49