Make jackson a test-only dependency
We are only using jackson to be able to test feeding data of the same
format into the python implementation and our own, such that we can
compare the results. We don't need jackson at run time, so split it
out. With the removal of xgboost4j and jackson moved to test-only this
package should no longer require dependencies to load into spark.

Change-Id: Ic48e711fd4381e7ba2f44378c1192831c9a76429