Convert es-ltr plugin integration to use elasticsearch-py
Implemented in two parts, a low level Client that exposes all of
the REST endpoints in the ltr plugin, and higher level DSL objects
for the entities that the plugin stores. These try to keep their
API's relatively similar to the elasticsearch-py and elasticsearch-dsl
packages respectively. Additionally this tries to follow roughly the
same abstractions as the elasticsearch plugin, using the same
implementation for the various object stores.

Minimal testing is done here, as I'm not sure how effectively we can
test this without having a running elasticsearch server. Perhaps in a
future update we can look at requiring an elasticsearch instance with
the plugin to be available at test time.

Change-Id: I47bf8fe8478d6bfce626510cb4b18b75d4ff95eb