Ensure hdfs files maintain correct ownership
When shelling out from a spark executor it appears we are writing
files out as the `yarn` user, rather than the user that ran the spark
job. It appears spark sources the user name separately and sets up
the appropriate context, to use the hadoop shell utils we need to
set that up ourselves.

Change-Id: I24d3ff581172e534ac57c7108eecda15697eba7d