bulk_daemon: Move assumption to action impl
The daemon was making a few assumptions about the content
that will be fetched from swift, but these assumptions don't
hold for a followup that implements ltr model upload. Move
the assumptions into the individual actions which are in a
better place to make these kinds of assumptions. While here
simplify Generator annotations into Iterator.

* Assumed all content was utf8, or decodable to utf8.
* Assumed files prefixed with '_' were to be ignored

Change-Id: Ida1c93a9801f3230bc0727372e88204d7a4d9f67