Revert "stream gzip decompression"
Since turning this on we've had problems with the bulk daemons getting
stuck (consuming but not completing a record from kafka). Stack traces
while stuck look like we are in the middle of streaming a response
from swift to elasticsearch. All threads are in a waiting state,
but no log message is emitted that the download completed. This
suggests that the download (not it's own thread, it's a generator
in the main execution thread) has not completed, but is no longer
emitting lines.

By reverting this patch we will no longer stream gzip files over
a long time period, the stream will be consumed immediately and
written to disk. There isn't any strong proof that the problem
is holding the stream open too long, but finishing the stream up
front will at least disentangle downloading from from swift
and uploading to elastic, to make clear which is the issue.

This reverts commit 5f4c5083e66af6dfb55e8ba2a23d9e07baa0da77.

Change-Id: I24293f8071048cd1aec5f909c1c103b94dc934d4