bulk_daemon: Remove parallelism from bulk import
The bulk daemons appear to be deadlocking. Stack traces while
the daemons are stuck indicates all threads waiting on either
a lock or a queue (or serving prometheus). Switch to the single
threaded bulk import implementation to verify this is the case.

While parallelism isn't a linear speedup, if it was this change would
mean we now need 2 weeks to import 1 week of updates. This is
a test, but cannot be the complete solution.

Change-Id: I508afb7680af0e9d92cfdde18be3351549cde98a