Revert "Revert "Revert "stream gzip decompression"""
This reverts commit e2f6ee9bb66c378e712928e0265d8c8adf50f4a8.

We redeployed this since it was unrelated to the incident which
triggered it's un-deployment. Since then we are logging occasional
EOFError's from gzip, where it says the stream is incomplete. This
patch's only purpose was to make things nicer, rather than digging into
the EOFError go back to the previous impl.

Comments around the related code have been updated in this revert to
reflect our new understanding.

Change-Id: I87a5f34832d25e9b148b28bb6288e6b39ddcd996